WIP: Changed floats to doubles in parameters

Closes #1971

See merge request orfeotoolbox/otb!623

+++ Modules/Wrappers/ApplicationEngine/include/otbWrapperApplication.h
@@ -421,7 +421,7 @@ public:
    * Can be called for types :
    * \li ParameterType_Float
    */
-  float GetParameterFloat(std::string const& parameter) const;
+  double GetParameterFloat(std::string const& parameter) const;
 
   /* Get a string parameter value
    *
